+++ src/app/main_tabs/tables/mod.rs
@@ -11,4 +11,5 @@ pub use heal_table::HealTable;
 pub use heal_table::HealTablePart;
 pub use heal_table::HealTablePartData;
 pub use metrics_table::TableSelectionEvent;
+pub use metrics_table::show_group_separator;
 pub use summary_table::SummaryTable;